+++ CCIA/Models/DistrictCounty.cs
@@ -5,9 +5,9 @@ namespace CCIA.Models
 {
     public partial class DistrictCounty
     {
-        public short DistCountyId { get; set; }
-        public short? DistId { get; set; }
-        public short? CountyId { get; set; }
+        public int DistCountyId { get; set; }
+        public int? DistId { get; set; }
+        public int? CountyId { get; set; }
         public string Comments { get; set; }
         public DateTime? DateEntered { get; set; }
         public string UserModified { get; set; }
